+++ packages/inngest/src/components/execution/v0.ts
@@ -200,13 +200,16 @@ export class V0InngestExecution
           .finally(() => {
             this.state.executingStep = false;
           })
-          .catch(async (error: Error) => {
-            return await this.transformOutput({ error }, outgoingUserFnOp);
-          })
-          .then(async (data) => {
-            await this.state.hooks?.afterExecution?.();
-            return await this.transformOutput({ data }, outgoingUserFnOp);
-          });
+          .then(
+            async (data) => {
+              await this.state.hooks?.afterExecution?.();
+              return await this.transformOutput({ data }, outgoingUserFnOp);
+            },
+            async (error: Error) => {
+              await this.state.hooks?.afterExecution?.();
+              return await this.transformOutput({ error }, outgoingUserFnOp);
+            },
+          );
 
         const { type: _type, ...rest } = result;
 
@@ -561,9 +564,18 @@ export class V0InngestExecution
        * Ensure we give middleware the chance to decide on retriable behaviour
        * by looking at the error returned from output transformation.
        */
-      let retriable: boolean | string = !(error instanceof NonRetriableError);
-      if (retriable && error instanceof RetryAfterError) {
-        retriable = error.retryAfter;
+      let retriable: boolean | string = !(
+        error instanceof NonRetriableError ||
+        // biome-ignore lint/suspicious/noExplicitAny: instanceof fails across module boundaries
+        (error as any)?.name === "NonRetriableError"
+      );
+      if (
+        retriable &&
+        (error instanceof RetryAfterError ||
+          // biome-ignore lint/suspicious/noExplicitAny: instanceof fails across module boundaries
+          (error as any)?.name === "RetryAfterError")
+      ) {
+        retriable = (error as RetryAfterError).retryAfter;
       }
 
       const serializedError = serializeError(error);