+++ packages/inngest/src/deno/fresh.test.ts
@@ -64,7 +64,8 @@ testFramework("Deno Fresh", DenoFreshHandler, {
     // biome-ignore lint/suspicious/noExplicitAny: intentional
     (req as any).headers = headers;
     // biome-ignore lint/suspicious/noExplicitAny: intentional
-    (req as any).json = () => Promise.resolve(req.body);
+    (req as any).text = () =>
+      Promise.resolve(req.body === undefined ? "" : JSON.stringify(req.body));
 
     return [req, env];
   },