+++ packages/inngest/src/helpers/consts.ts
@@ -151,6 +151,7 @@ export enum headerKeys {
   TraceState = "tracestate",
   InngestRunId = "x-run-id",
   InngestStepId = "x-inngest-step-id",
+  InngestForceStepPlan = "x-inngest-force-step-plan",
 }
 
 /**
@@ -179,16 +180,7 @@ export enum internalEvents {
   FunctionFinished = "inngest/function.finished",
   FunctionCancelled = "inngest/function.cancelled",
   ScheduledTimer = "inngest/scheduled.timer",
-}
-
-/**
- * Events that are known globally by the Inngest platform.
- */
-export enum knownEvents {
-  /**
-   * An HTTP request has been received to trigger a function execution.
-   */
-  HttpRunStarted = "http/run.started",
+  HttpRequest = "inngest/http.request",
 }
 
 export const logPrefix: string = chalk.magenta.bold("[Inngest]");